// frontend/angular/projects/apicat-app/src/model/utenteOrganizzazione.ts
/*
 * GovCat - GovWay API Catalogue
 * https://github.com/link-it/govcat
 *
 * Copyright (c) 2021-2026 Link.it srl (https://link.it).
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License version 3, as published by
 * the Free Software Foundation.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */

/**
 * MODI - APICatalogo
 * Registrazione e Adesione alle API
 *
 * The version of the OpenAPI document: 1.5.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ItemOrganizzazione } from './itemOrganizzazione';
import { RuoloOrganizzazioneEnum } from './ruoloOrganizzazioneEnum';


export interface UtenteOrganizzazione {
    organizzazione: ItemOrganizzazione;
    /**
     * Ruolo dell'utente nell'organizzazione. `null` significa "nessun
     * ruolo" (utente associato in sola lettura).
     */
    ruolo_organizzazione?: RuoloOrganizzazioneEnum | null;
}